requirements.txt
# Core dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langgraph==0.0.20
pydantic==2.5.0

# LLM and embeddings
transformers==4.35.2
torch==2.4.1
sentence-transformers==2.2.2
huggingface_hub==0.25.2  # Compatible with sentence-transformers 2.2.2
accelerate==0.25.0
einops  # Required for MPT models

# Document processing
pypdf==3.17.1
python-docx==1.1.0
markdown==3.5.1
beautifulsoup4==4.12.2

# Utilities
python-dotenv==1.0.0
tiktoken==0.5.2
chromadb==0.4.18

# CLI and interface
streamlit==1.28.2
rich==13.7.0

# Optional: for better performance
protobuf==4.25.1
# Note: sentencepiece is optional and may require build tools on macOS.
# Install separately if needed for specific models: pip install sentencepiece

